Rename data-provier to symbols-table
