+++ ts/editor/symbols-overlay/symbols-table.ts
